fix(settings): extract SettingsRow to its own module
Next.js page files cannot export arbitrary symbols — the build failed with
"SettingsRow is not a valid Page export field". Move SettingsRow +
SettingsItem into ./SettingsRow.tsx and re-import them in page.tsx so the
test can still reach SettingsRow directly without breaking Next's page
export contract.
